Removing the cost function creation code, experimental data, and pool creation from NS class. My thoughts were that this way there is a separation of trainer and model.
Just for my own sake, I renamed functions for running and saving output (run & save), so I can call them multiple times. It might be useful to continue a NS run if you ran it for so long, then you can load in the saved state and restart? Just a thought.

For the run file, I added an objective function using pandas. The exact costfunction is similar to the EARM2 paper, without the spline fitting for smac, but it is easily adaptable. I just wanted to move the cost function out of the class and didnt think I could figure out the other cost function code without breaking it.